fix: init sign-in for login screen
