--- frontend/src/lib/proxy/$public/app.services.proxy.ts
@@ -1,8 +1,0 @@
-const importAppServices = () => import("../../services/$public/app.services");
-
-export const initAppProxy = async (): Promise<
-  [PromiseSettledResult<void[]>, PromiseSettledResult<void[]>]
-> => {
-  const { initApp } = await importAppServices();
-  return initApp();
-};
