+++ frontend/src/routes/(login)/+layout.svelte
@@ -1,5 +1,9 @@
 <script lang="ts">
   import Banner from "$lib/components/header/Banner.svelte";
+  import { onMount } from "svelte";
+  import { initAppSignIn } from "$lib/services/$public/app.services";
+
+  onMount(async () => await initAppSignIn());
 </script>
 
 <div class="container" data-tid="auth-layout">
